Fix getInitialPath to use props.history instead of props.useHistory
Uncaught Error: No route matched path: / when used on a page in a
subdirectory.
The documentation states { history : true } should be passed to the
component which attaches it to the component’s props. useHistory
becomes a property of the component state which isn’t available yet
when getInitialPath is called.